Reorganizes directory structure
