+++ src/store/configureStore.dev.js
@@ -2,7 +2,7 @@ import { routerMiddleware } from 'react-router-redux';
 import { createStore, compose, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
-import root from '../reducers/';
+import root from './rootReducer';
 
 export default function configureStore(history) {
   const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ||
@@ -14,8 +14,8 @@ export default function configureStore(history) {
 
   if (module.hot) {
     // Enable Webpack hot module replacement for reducers
-    module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers').default;
+    module.hot.accept('./rootReducer', () => {
+      const nextRootReducer = require('./rootReducer').default;
       store.replaceReducer(nextRootReducer);
     });
   }
