+++ src/store/configureStore.prod.js
@@ -3,7 +3,7 @@ import { routerMiddleware } from 'react-router-redux';
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
-import root from '../reducers/';
+import root from './rootReducer';
 
 export default function configureStore() {
   return createStore(
